relocate timestamp_to_index
