+++ PIconnect/time.py
@@ -1,6 +1,8 @@
 from datetime import datetime
+import pytz
 
 from PIconnect.AFSDK import AF
+from PIconnect.config import PIConfig
 
 
 def to_af_time_range(start_time, end_time):
@@ -19,3 +21,29 @@ def to_af_time_range(start_time, end_time):
         end_time = end_time.isoformat()
 
     return AF.Time.AFTimeRange(start_time, end_time)
+
+
+def timestamp_to_index(timestamp):
+    """Convert AFTime object to datetime in local timezone.
+
+    .. todo::
+
+        Allow to define timezone, default to UTC?
+
+    .. todo::
+
+    """
+    local_tz = pytz.timezone(PIConfig.DEFAULT_TIMEZONE)
+    return (
+        datetime(
+            timestamp.Year,
+            timestamp.Month,
+            timestamp.Day,
+            timestamp.Hour,
+            timestamp.Minute,
+            timestamp.Second,
+            timestamp.Millisecond * 1000,
+        )
+        .replace(tzinfo=pytz.utc)
+        .astimezone(local_tz)
+    )
